Move LoginContent to components folder, and expose it as entry point wrapped in a Suspense
